Llama-hybrid-FAISS-BM25--rag-assistant/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
nltk>=3.9.1
pyarrow>=16.1.0
duckdb>=1.1.2
# LlamaIndex core + OpenAI HuggingFace integrations (optional; UIanswering can work without)
llama-index-core>=0.11.20
llama-index-embeddings-huggingface>=0.3.1
llama-index-llms-openai>=0.3.5
# Streamlit chat UI
streamlit==1.39.0